requirements.txt
pymongo
bcrypt
pyjwt
